+++ BE/src/main/java/com/example/StudyBoard/application/repository/ApplicationRepository.java
@@ -1,6 +1,6 @@
-package com.example.StudyBoard.repository;
+package com.example.StudyBoard.application.repository;
 
-import com.example.StudyBoard.entity.Application;
+import com.example.StudyBoard.application.entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
