+++ BE/src/main/java/com/example/StudyBoard/auth/entity/CustomMemberDetails.java
@@ -1,6 +1,6 @@
 package com.example.StudyBoard.auth.entity;
 
-import com.example.StudyBoard.entity.Member;
+import com.example.StudyBoard.member.entity.Member;
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
