+++ BE/src/main/java/com/example/StudyBoard/auth/service/AuthService.java
@@ -3,10 +3,10 @@ package com.example.StudyBoard.auth.service;
 import com.example.StudyBoard.auth.dto.*;
 import com.example.StudyBoard.auth.entity.RefreshToken;
 import com.example.StudyBoard.auth.repository.TokenRepository;
-import com.example.StudyBoard.entity.Member;
+import com.example.StudyBoard.member.entity.Member;
 import com.example.StudyBoard.exception.BusinessException;
 import com.example.StudyBoard.exception.ErrorCode;
-import com.example.StudyBoard.repository.MemberRepository;
+import com.example.StudyBoard.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
